Add support for specifying part sizes and stdin as input
This allows piping directly to tarsplitter from tar itself when input is
specified as '-', which avoids the need for a large intermediary tar file.
For example: `tar -cvf - . | tarsplitter -i - -s 1G -o /tmp/archive-`.
This will create tar files that are at most 1GB in size, though
individual sizes will vary depending on the input files and how they're
sorted.

The `-p` option doesn't make sense when input is stdin since we can't
know the total input size in advance to calculate the part size.
Similarly, if `-s` is not provided when input is stdin no splitting will
occur and in both cases only a single tar file will be created, which
defeats the purpose of using tarsplitter, so make sure to always specify
`-s` with `-i -`.